update hello world docs
